Extend Golang's OpenPGP library with some custom functions
Golang's OpenPGP library implementation does not provide all the APIs we
need for image encryption. In particular, we want to be able to add
recipients to an image by encrypting the symmetric key with which a
layer is encrypted with the public key of a new recipient. Further, we
need to be able to decrypt the symmetric key in a separate step so that
other technolgies, such as hardware security modules (HSMs) or TPMs, can
be used to release a symmetric key. The existing Golang OpenPGP library
has a high level API where one would feed in an encrypted bystream and
it would decrypt the symmetric key using the private key of a user and
then also decrypt the payload. With our API extensions we split this
into two separate parts of symmetric key decryption and payload de-
cryption.

Signed-off-by: Stefan Berger <stefanb@linux.vnet.ibm.com>
Signed-off-by: Brandon Lum <lumjjb@gmail.com>