Implement CheckAuthorization() to check whether user may use image
Implement CheckAuthorization() that checks whether a user has the right keys
to every layer of the current platform so that he would be authorized to
create a container from the image. We do this by using the provided keys
for unwrapping of the layer encryption key and return no error if it
succeeds and an error message otherwise.

Signed-off-by: Stefan Berger <stefanb@linux.ibm.com>