ctr: Remove stale snapshot in case of failure to create a container
If the creation of a container fails it may leave behind a stale snapshot
that needs to be deleted before retrying to create the container. We clean
this stale snapshot up after a failed attempt to create a container.

To avoid deleting a snapshot in the case a container with a name of an
existing container is created, we make sure that no container with the
same name already exists. The reason is that deleting a stale snapshot
while not deleting the container would result in problems when trying
to delete the container.

Also, we make sure that a snapshot with the same name does not already
exist so we do not accidentally delete a pre-existing snapshot when the
container creation failed.

Signed-off-by: Stefan Berger <stefanb@linux.ibm.com>