Pass ReaderAt into encryption function instead of plain byte array
This is a first step in the conversion of using the notion of
message streams in the block cipher implementation. In this patch
we convert the code to pass a content.ReadAt to the encryption
function so that later on we can replace the symmetric cipher
with an implementation that works on blocks read from this reader.

Signed-off-by: Stefan Berger <stefanb@linux.ibm.com>