Refactor GCM block cipher to do all work in the Reader
Refactor the GCM block cipher to do all the work in the Reader's Read()
function. This changes the semantics of the Size() function a bit, which
will only return the size of the encrypted or decrypted data after
Read() has been called for all the data. So the test cases need to be
adjusted for this. Also errors related to decrypting the data with
the wrong key only become apparent when calling Read(), which also
requires test cases to be adapted.

Signed-off-by: Stefan Berger <stefanb@linux.ibm.com>