Use AES-SIV-CMAC-512 rather than AEAD GCM
Use AES-SIV-CMAC-512 rather than AEAD GCM and get rid of the latter.

The AES-SIV-CMAC-512 is used by reading chunks from the input stream,
encrypting or decrypting the chunk, and then returning the data through
the buffer passed to the Read() function.

When encrypting a chunk the size of the encrypted chunk is a certain amount
larger than the plain one. We have to make sure that the encrypted chunks
passed back to the caller are given to us by the caller exactly the same
when decrypting them. We make the chunksize of an encrypted chunk 1MB but
only read plain data of the size 1MB - Overhead.

Signed-off-by: Stefan Berger <stefanb@linux.ibm.com>